move Server object out of utils and clean up
