+++ .env.example
@@ -26,4 +26,6 @@ MANAGEMENT_ROLE_ID=
 
 UPVOTE_EMOJI_ID=
 DOWNVOTE_EMOJI_ID=
+
 NOSTARBOARD_EMOJI_ID=
+STARBOARD_AMOUNT=
