--- src/main/kotlin/com/learnspigot/bot/help/OpenThreadListener.kt
@@ -1,9 +1,0 @@
-package com.learnspigot.bot.help
-
-import net.dv8tion.jda.api.events.thread.GenericThreadEvent
-import net.dv8tion.jda.api.events.thread.ThreadRevealedEvent
-import net.dv8tion.jda.api.hooks.ListenerAdapter
-
-class OpenThreadListener : ListenerAdapter() {
-
-}
