+++ src/main/kotlin/com/learnspigot/bot/knowledgebase/KnowledgebaseListener.kt
@@ -1,7 +1,7 @@
 package com.learnspigot.bot.knowledgebase
 
 import com.learnspigot.bot.Environment
-import com.learnspigot.bot.util.Server
+import com.learnspigot.bot.Server
 import net.dv8tion.jda.api.entities.channel.ChannelType
 import net.dv8tion.jda.api.events.channel.update.ChannelUpdateArchivedEvent
 import net.dv8tion.jda.api.hooks.ListenerAdapter
