+++ src/main/kotlin/com/learnspigot/bot/reputation/LeaderboardMessage.kt
@@ -1,8 +1,7 @@
 package com.learnspigot.bot.reputation
 
-import com.learnspigot.bot.Environment
 import com.learnspigot.bot.profile.ProfileRegistry
-import com.learnspigot.bot.util.Server
+import com.learnspigot.bot.Server
 import com.learnspigot.bot.util.embed
 import net.dv8tion.jda.api.entities.Guild
 import net.dv8tion.jda.api.entities.MessageEmbed
