+++ src/main/kotlin/com/learnspigot/bot/starboard/StarboardEntry.kt
@@ -1,6 +1,5 @@
 package com.learnspigot.bot.starboard
 
-import com.learnspigot.bot.util.Server
 import org.bson.Document
