+++ src/main/kotlin/com/learnspigot/bot/suggestion/SuggestionListener.kt
@@ -1,10 +1,9 @@
 package com.learnspigot.bot.suggestion
 
 import com.learnspigot.bot.Environment
-import com.learnspigot.bot.util.Server
+import com.learnspigot.bot.Server
 import com.learnspigot.bot.util.embed
 import net.dv8tion.jda.api.entities.Message
-import net.dv8tion.jda.api.entities.emoji.Emoji
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 import net.dv8tion.jda.api.hooks.ListenerAdapter
 import java.util.concurrent.TimeUnit
